% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.annotation.R
\name{as_df.annotation}
\alias{as_df.annotation}
\title{Create a dataframe with annotation content}
\usage{
as_df.annotation(sbml_list_listOf)
}
\arguments{
\item{sbml_list_listOf}{sbml-converted list restricted to a 'listOf' level (e.g. listOfSpecies), a list}
}
\value{
one data frame with one row for each entity inside the 'listOf' selected and one column for each annotation's descriptor reported for them.  Empty vector if the 'annotation' content is empty for each entity.
}
\description{
It returns a dataframe with one row for each entity and 0 or more columns. If zero, it means the annotation content for each entity inside the 'listOf' selected is empty.
}
\details{
Each column is named as 'annotation_' concatenated with the 'descriptor name' (e.g. is, hasPart, isVersionOf). If the name is repeated (i.e. there are multiple equal descriptors' name for the same entity), it concatenates also a number to the name, starting from '_1' and leaving the first name without any number.
}
\examples{
sbml_list <- sbml_as_list(system.file("extdata", "R-HSA-8937144.sbml", package = "tidysbml"), "species")
df <- as_df.annotation(sbml_list)  ## returns one dataframe with one row for each species and one column for each descriptor in annotation
}
