% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.attributes.R
\name{as_df.attributes}
\alias{as_df.attributes}
\title{Create a dataframe with attributes content}
\usage{
as_df.attributes(sbml_list_listOf)
}
\arguments{
\item{sbml_list_listOf}{sbml converted into a list restricted to a 'listOf' level (e.g. listOfSpecies), a list}
}
\value{
a data frame with one row for each entity inside the 'listOf' chosen and one column for each attribute reported for them
}
\description{
It returns a dataframe containing information about attributes for each entity reported in the 'listOf' class selected. It has at least one column. Otherwise, the format of the input list is not appropriate for this type of extraction.
}
\details{
The input list may be obtained using sbml_as_list() function, by selecting as second argument the type of entity to be selected (i.e. either 'species', 'reactions' or 'compartments').
}
\examples{
sbml_list <- sbml_as_list(system.file("extdata", "R-HSA-8937144.sbml", package = "tidysbml"), "species")
df <- as_df.attributes(sbml_list)   ## returns one dataframe with one row for each species and one column for each attribute
}
