% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.notes.R
\name{as_df.notes}
\alias{as_df.notes}
\title{Create a dataframe with notes content}
\usage{
as_df.notes(sbml_list_listOf)
}
\arguments{
\item{sbml_list_listOf}{sbml-converted list restricted to a 'listOf' level (e.g. listOfSpecies), a list}
}
\value{
a data frame with one row for each entity inside the 'listOf' chosen and one column for the 'notes' content. Empty vector if the 'notes' content is empty for each entity.
}
\description{
It returns a dataframe with either zero or one column. It is zero if the notes content is empty for all the entities inside the 'listOf' class selected.
}
\examples{
sbml_list <- sbml_as_list(system.file("extdata", "R-HSA-8937144.sbml", package = "tidysbml"), "species")
df <- as_df.notes(sbml_list)  ## returns the dataframe with only one column containing the notes' info
}
