% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_subdf.R
\name{as_subdf}
\alias{as_subdf}
\title{Create a dataframe for speciesReference entities}
\usage{
as_subdf(sbml_list_listOf)
}
\arguments{
\item{sbml_list_listOf}{sbml-converted list restricted to the 'listOfReactions' level, a list}
}
\value{
a data frame with one row for each speciesReference entity and one column for each attribute, notes, annotation reported for them
}
\description{
It returns a dataframe containing data (i.e. attributes, notes and annotation) about the species involved in each reaction described in listOfReactions.
}
\details{
It will create the data frame if and only if listOfReactions is present within the sbml in argument and is not empty (i.e. it contains 'speciesReference' entities).
The output dataframe has at least two columns, one with the id for the reaction taken into account, one with the type of container the species belongs to (i.e. listOfReactants, listOfProduct, listOfModifiers) and others for the mandatory attributes.
Otherwise, the format of the input list is not appropriate for this type of extraction.
}
\examples{
sbml_list <- sbml_as_list(system.file("extdata", "R-HSA-8937144.sbml", package = "tidysbml"), "reactions")
df <- as_subdf(sbml_list)   ## returns one dataframe with one row for each species involved in each reaction
}
